+++ packages/std/src/colors/colorDesaturate.ts
@@ -22,7 +22,7 @@ export default function colorDesaturate(
   amount: number,
   format: ColorFormat = 'hex',
 ): ColorOutput {
-  const hsla = color(input, 'hsla-object');
+  const hsla = color({ value: input, to: 'hsla-object' });
   hsla.s = Math.max(0, hsla.s - amount * 100);
-  return color(hsla, format);
+  return color({ value: hsla, to: format });
 }