+++ packages/std/src/colors/colorLighten.ts
@@ -22,7 +22,7 @@ export default function colorLighten(
   amount: number,
   format: ColorFormat = 'hex',
 ): ColorOutput {
-  const hsla = color(input, 'hsla-object');
+  const hsla = color({ value: input, to: 'hsla-object' });
   hsla.l = Math.min(100, hsla.l + amount * 100);
-  return color(hsla, format);
+  return color({ value: hsla, to: format });
 }